Widen lessonNumber Int→Float so chapter tests can number as 7.1
Chapter tests are authored as lesson-format markdown numbered X.1 and
must sort after lesson X. Both live schema fields were integers
(LessonTemplate.lessonNumber, Lesson.order), so the parser's parseInt
turned "7.1" into 7 — tying with Lesson 7 instead of following it.

- amplify/data/resource.ts: a.integer() → a.float() on both fields.
  Int→Float is a widening: every stored value is already an integer
  (Int! made anything else impossible to write), DynamoDB stores both
  GraphQL types as the same Number type, and the deploy touches only
  the AppSync schema, never the tables — no data migration needed.
- parseFloat replaces parseInt at every lessonNumber write path:
  markdown parser (regex now captures the decimal), library editor
  save, schedule page Lesson.order, and the video upload page (a
  fourth site the investigation report had missed).
- All ordering already compares numerically (a - b, verified
  repo-wide) — "7.1 after 70" string-sort bugs cannot occur.
- CLAUDE.md's Lesson Numbers section previously claimed Float storage
  while the schema said Int!; it now describes the actual state.
- Verified: tsc --noEmit clean; parser fixtures assert 7.1 → 7.1,
  integer 7 → 7 (number), "1.5a" → 1.5.

Also checks in claude/test-consumption-report.md, the read-only
investigation this change implements recommendation 1 of.

+++ app/teacher/library/[courseId]/page.tsx
@@ -520,7 +520,7 @@ export default function LessonLibraryPage() {
           input: {
             id,
             title: editForm.title,
-            lessonNumber: parseInt(editForm.lessonNumber) || 0,
+            lessonNumber: parseFloat(editForm.lessonNumber) || 0,
             instructions: editForm.instructions || null,
             teachingNotes: editForm.teachingNotes || null,
             worksheetUrl: editForm.worksheetUrl || null,
@@ -533,7 +533,7 @@ export default function LessonLibraryPage() {
       setLessons(prev => prev.map(l => l.id === id ? {
         ...l,
         title: editForm.title,
-        lessonNumber: parseInt(editForm.lessonNumber) || l.lessonNumber,
+        lessonNumber: parseFloat(editForm.lessonNumber) || l.lessonNumber,
         instructions: editForm.instructions || null,
         teachingNotes: editForm.teachingNotes || null,
         worksheetUrl: editForm.worksheetUrl || null,
